requirements.txt
Flask==2.3.2
Flask-SocketIO==5.3.4
PyYAML
aiohttp
requests
Flask-CORS
speedtest-cli